+++ src/PluginSettingsManager.ts
@@ -6,6 +6,7 @@ import { appendCodeBlock } from 'obsidian-dev-utils/HTMLElement';
 import { alert } from 'obsidian-dev-utils/obsidian/Modals/Alert';
 import { PluginSettingsManagerBase } from 'obsidian-dev-utils/obsidian/Plugin/PluginSettingsManagerBase';
 import { EmptyAttachmentFolderBehavior } from 'obsidian-dev-utils/obsidian/RenameDeleteHandler';
+import { getOsUnsafePathCharsRegExp } from 'obsidian-dev-utils/obsidian/Validation';
 import { isValidRegExp } from 'obsidian-dev-utils/RegExp';
 import { replaceAll } from 'obsidian-dev-utils/String';
 import { compare } from 'semver';
@@ -17,7 +18,6 @@ import {
   PluginSettings
 } from './PluginSettings.ts';
 import {
-  INVALID_FILENAME_PATH_CHARS_REG_EXP,
   parseCustomTokens,
   TokenValidationMode,
   validateFileName,
@@ -165,6 +165,29 @@ ${commentOut(legacySettings.customTokensStr)}
           : CollectAttachmentUsedByMultipleNotesMode.Skip;
       }
 
+      if (
+        legacySettings.warningVersion && compare(legacySettings.warningVersion, '9.2.0') < 0
+        // eslint-disable-next-line no-template-curly-in-string
+        && (legacySettings.markdownUrlFormat === '${generatedAttachmentFilePath}' || legacySettings.markdownUrlFormat === '${noteFilePath}')
+      ) {
+        invokeAsyncSafely(async () => {
+          await this.plugin.waitForLifecycleEvent('layoutReady');
+          await alert({
+            app: this.app,
+            message: createFragment((f) => {
+              f.appendText('You have potentially incorrect value set for the ');
+              appendCodeBlock(f, 'Markdown URL format');
+              f.appendText(' setting. Please refer to the ');
+              f.createEl('a', {
+                href: 'https://github.com/RainCat1998/obsidian-custom-attachment-location?tab=readme-ov-file#markdown-url-format',
+                text: 'documentation'
+              });
+              f.appendText(' for more information. This message will not be shown again.');
+            })
+          });
+        });
+      }
+
       legacySettings.warningVersion = this.plugin.manifest.version;
     });
   }
@@ -191,7 +214,7 @@ ${commentOut(legacySettings.customTokensStr)}
     });
 
     this.registerValidator('specialCharactersReplacement', (value): MaybeReturn<string> => {
-      if (INVALID_FILENAME_PATH_CHARS_REG_EXP.exec(value)) {
+      if (getOsUnsafePathCharsRegExp().exec(value)) {
         return 'Special character replacement must not contain invalid file name path characters.';
       }
     });